spam upload a file on server two times
first time the file is uploaded in file widgets/upload.js, a function
wich handle a upload progress bar. Then, when a user send form data to
server, the last file aviable in uploader (http input type file
element) a sendet again to the server.
this patch is a quick fix that can not work with multiple files upload
need mor tests.